// src/api/openapi/auto-generated/models/ResultSetDataQueryParams.ts
/* tslint:disable */
/* eslint-disable */ // @ts-nocheck
/**
 * Jupiter API
 * Welcome to Jupiter\'s ClimateScore Global API
 *
 * The version of the OpenAPI document: 2.2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Filters } from './Filters';
import {
    FiltersFromJSON,
    FiltersFromJSONTyped,
    FiltersToJSON,
} from './Filters';

/**
 * 
 * @export
 * @interface ResultSetDataQueryParams
 */
export interface ResultSetDataQueryParams {
    /**
     * 
     * @type {Array<number>}
     * @memberof ResultSetDataQueryParams
     */
    locationIds?: Array<number>;
    /**
     * 
     * @type {Array<string>}
     * @memberof ResultSetDataQueryParams
     */
    include?: Array<string>;
    /**
     * 
     * @type {{ [key: string]: Filters; }}
     * @memberof ResultSetDataQueryParams
     */
    filters?: { [key: string]: Filters; };
}

/**
 * Check if a given object implements the ResultSetDataQueryParams interface.
 */
export function instanceOfResultSetDataQueryParams(value: object): value is ResultSetDataQueryParams {
    return true;
}

export function ResultSetDataQueryParamsFromJSON(json: any): ResultSetDataQueryParams {
    return ResultSetDataQueryParamsFromJSONTyped(json, false);
}

export function ResultSetDataQueryParamsFromJSONTyped(json: any, ignoreDiscriminator: boolean): ResultSetDataQueryParams {
    if (json == null) {
        return json;
    }
    return {
        
        'locationIds': json['location_ids'] == null ? undefined : json['location_ids'],
        'include': json['include'] == null ? undefined : json['include'],
        'filters': json['filters'] == null ? undefined : (mapValues(json['filters'], FiltersFromJSON)),
    };
}

export function ResultSetDataQueryParamsToJSON(value?: ResultSetDataQueryParams | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'location_ids': value['locationIds'],
        'include': value['include'],
        'filters': value['filters'] == null ? undefined : (mapValues(value['filters'], FiltersToJSON)),
    };
}