// src/api/openapi/auto-generated/models/ResultSetDataSchema.ts
/* tslint:disable */
/* eslint-disable */ // @ts-nocheck
/**
 * Jupiter API
 * Welcome to Jupiter\'s ClimateScore Global API
 *
 * The version of the OpenAPI document: 2.2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ResultSetDataSchema
 */
export interface ResultSetDataSchema {
    /**
     * 
     * @type {string}
     * @memberof ResultSetDataSchema
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof ResultSetDataSchema
     */
    enhancedName: string;
    /**
     * 
     * @type {string}
     * @memberof ResultSetDataSchema
     */
    userGuideDefinition: string;
    /**
     * 
     * @type {string}
     * @memberof ResultSetDataSchema
     */
    peril?: ResultSetDataSchemaPerilEnum;
}


/**
 * @export
 */
export const ResultSetDataSchemaPerilEnum = {
    cold: 'cold',
    combinedFlood: 'combinedFlood',
    drought: 'drought',
    fire: 'fire',
    hail: 'hail',
    heat: 'heat',
    precip: 'precip',
    wind: 'wind',
    coastalFlood: 'coastalFlood',
    fluvialFlood: 'fluvialFlood',
    pluvialFlood: 'pluvialFlood',
    unknownDefaultOpenApi: '11184809'
} as const;
export type ResultSetDataSchemaPerilEnum = typeof ResultSetDataSchemaPerilEnum[keyof typeof ResultSetDataSchemaPerilEnum];


/**
 * Check if a given object implements the ResultSetDataSchema interface.
 */
export function instanceOfResultSetDataSchema(value: object): value is ResultSetDataSchema {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('enhancedName' in value) || value['enhancedName'] === undefined) return false;
    if (!('userGuideDefinition' in value) || value['userGuideDefinition'] === undefined) return false;
    return true;
}

export function ResultSetDataSchemaFromJSON(json: any): ResultSetDataSchema {
    return ResultSetDataSchemaFromJSONTyped(json, false);
}

export function ResultSetDataSchemaFromJSONTyped(json: any, ignoreDiscriminator: boolean): ResultSetDataSchema {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'enhancedName': json['enhanced_name'],
        'userGuideDefinition': json['user_guide_definition'],
        'peril': json['peril'] == null ? undefined : json['peril'],
    };
}

export function ResultSetDataSchemaToJSON(value?: ResultSetDataSchema | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'id': value['id'],
        'enhanced_name': value['enhancedName'],
        'user_guide_definition': value['userGuideDefinition'],
        'peril': value['peril'],
    };
}