// src/types/locationDamageAndLossData.ts
export class LocationDamageAndLossData {
    FL_avgAnnualLoss_total?: number | null;
    totalValue?: number | null;
    FL_avgAnnualDamage_building?: number | null;
    FL_avgAnnualDamage_contents?: number | null;
    FL_avgAnnualDamage_inventory?: number | null;
    FL_avgAnnualDamage_downtime?: number | null;
    WS_avgAnnualLoss_total?: number | null;
    WS_avgAnnualDamage_building?: number | null;
    WS_avgAnnualDamage_contents?: number | null;
    WS_avgAnnualDamage_inventory?: number | null;
    WS_avgAnnualDamage_downtime?: number | null;
    FR_avgAnnualLoss_total?: number | null;
    FR_avgAnnualDamage_building?: number | null;
    FR_avgAnnualDamage_contents?: number | null;
    FR_avgAnnualDamage_inventory?: number | null;
    FR_avgAnnualDamage_downtime?: number | null;
    HT_avgAnnualDamage_cooling?: number | null;
    HT_avgAnnualDamage_prod_abs?: number | null;
    DT_shadowPriceOfWater?: number | null;
    FL_avgAnnualLoss_building?: number | null;
    FL_avgAnnualLoss_contents?: number | null;
    FL_avgAnnualLoss_inventory?: number | null;
    FL_avgAnnualLoss_downtime?: number | null;
    WS_avgAnnualLoss_building?: number | null;
    WS_avgAnnualLoss_contents?: number | null;
    WS_avgAnnualLoss_inventory?: number | null;
    WS_avgAnnualLoss_downtime?: number | null;
    FR_avgAnnualLoss_building?: number | null;
    FR_avgAnnualLoss_contents?: number | null;
    FR_avgAnnualLoss_inventory?: number | null;
    FR_avgAnnualLoss_downtime?: number | null;
    HT_avgAnnualLoss_cooling?: number | null;
    HT_avgAnnualLoss_prod?: number | null;
    DT_avgAnnualLoss_water?: number | null;
    locationId?: number;
    year?: number;
    valueCurrencyCode?: string;
}