// src/types/locationFinancialMetricsData.ts
export class LocationFinancialMetricsData {
    totalValue?: number | null;
    locationId?: number;
    year?: number;
    valueCurrencyCode?: string;
    EI_netIncome?: number;
    EI_technicalPremium?: number;
    EI_benchmarkValue?: number;
    EI_climateAdjValue?: number;
    EI_value?: number;
    EI_overheads?: number;
    EI_productivity?: number;
    EI_inventory?: number;
    EI_downtime?: number;
    EI_building?: number;
    EI_contents?: number;
    EI_ownershipCost?: number;
}